Restructure examples
